scripts/train.py
import yaml
import ray
from models import TemporalFusion, HierarchicalAgent
from utils.replay_buffer import DualPriorityReplayBuffer

@ray.remote(num_gpus=0.5)
class Learner:
    def __init__(self, config):
        self.config = config
        self.temporal_model = TemporalFusion()
        self.agents = [HierarchicalAgent(agent_type=t) 
                      for t in ['leaf', 'rack', 'cluster']]
        self.replay = DualPriorityReplayBuffer()
        
    def update(self, batch):
        # Distributed training logic here
        pass

def main():
    with open('configs/default.yaml') as f:
        config = yaml.safe_load(f)
    
    ray.init()
    
    # Create 8 learners and 64 actors
    learners = [Learner.remote(config) for _ in range(8)]
    actors = [DataActor.remote(config) for _ in range(64)]
    
    # Training loop
    for epoch in range(config['epochs']):
        # Distributed training logic
        pass

if __name__ == "__main__":
    main()